packages/robot-agent-py/pyproject.toml
[project]
name = "robocloud-agent"
version = "0.1.0"
description = "Python robot agent for RoboCloud — parity implementation of the TypeScript robot-agent-ts"
requires-python = ">=3.10"
dependencies = [
    "mujoco>=3.1",
    "numpy>=1.24",
    "websockets>=12.0",
    "protobuf>=5.0",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
robocloud-agent = "robocloud_agent.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
